CTC segmentation - apply feedback
- fix the stage numbering in run.sh

- separate ESPnet- and segmentation-specific functions (for CTC segmentation as separate package)
